+++ src/sql/magic.py
@@ -21,7 +21,7 @@ import warnings
 import sql.connection
 import sql.parse
 import sql.run
-from sql import display
+from sql import display, exceptions
 from sql.store import store
 from sql.command import SQLCommand
 from sql.magic_plot import SqlPlotMagic
@@ -29,12 +29,13 @@ from sql.magic_cmd import SqlCmdMagic
 from sql._patch import patch_ipython_usage_error
 from ploomber_core.dependencies import check_installed
 
+from sql.error_message import detail
 from traitlets.config.configurable import Configurable
-from traitlets import Bool, Int, Unicode, Dict, observe
+from traitlets import Bool, Int, TraitError, Unicode, Dict, observe, validate
 
 try:
     from pandas.core.frame import DataFrame, Series
-except ImportError:
+except ModuleNotFoundError:
     DataFrame = None
     Series = None
 
@@ -94,7 +95,7 @@ class SqlMagic(Magics, Configurable):
         help="Don't display the full traceback on SQL Programming Error",
     )
     displaylimit = Int(
-        None,
+        sql.run.DEFAULT_DISPLAYLIMIT_VALUE,
         config=True,
         allow_none=True,
         help=(
@@ -144,6 +145,26 @@ class SqlMagic(Magics, Configurable):
         # Add ourself to the list of module configurable via %config
         self.shell.configurables.append(self)
 
+    # To verify displaylimit is valid positive integer
+    # If:
+    #   None -> We treat it as 0 (no limit)
+    #   Positive Integer -> Pass
+    #   Negative Integer -> raise Error
+    @validate("displaylimit")
+    def _valid_displaylimit(self, proposal):
+        if proposal["value"] is None:
+            print("displaylimit: Value None will be treated as 0 (no limit)")
+            return 0
+        try:
+            value = int(proposal["value"])
+            if value < 0:
+                raise TraitError(
+                    "{}: displaylimit cannot be a negative integer".format(value)
+                )
+            return value
+        except ValueError:
+            raise TraitError("{}: displaylimit is not an integer".format(value))
+
     @observe("autopandas", "autopolars")
     def _mutex_autopandas_autopolars(self, change):
         # When enabling autopandas or autopolars, automatically disable the
@@ -180,6 +201,12 @@ class SqlMagic(Magics, Configurable):
         "--persist",
         action="store_true",
         help="create a table name in the database from the named DataFrame",
+    )
+    @argument(
+        "-P",
+        "--persist-replace",
+        action="store_true",
+        help="replace the DataFrame if it exists, otherwise perform --persist",
     )
     @argument(
         "-n",
@@ -261,6 +288,7 @@ class SqlMagic(Magics, Configurable):
         )
 
     @telemetry.log_call("execute", payload=True)
+    @modify_exceptions
     def _execute(self, payload, line, cell, local_ns, is_interactive_mode=False):
         def interactive_execute_wrapper(**kwargs):
             for key, value in kwargs.items():
@@ -346,11 +374,34 @@ class SqlMagic(Magics, Configurable):
         )
         payload["connection_info"] = conn._get_curr_sqlalchemy_connection_info()
 
-        if args.persist:
+        if args.persist_replace and args.append:
+            raise exceptions.UsageError(
+                """You cannot simultaneously persist and append data to a dataframe;
+                  please choose to utilize either one or the other."""
+            )
+        if args.persist and args.persist_replace:
+            warnings.warn("Please use either --persist or --persist-replace")
+            return self._persist_dataframe(
+                command.sql,
+                conn,
+                user_ns,
+                append=False,
+                index=not args.no_index,
+                replace=True,
+            )
+        elif args.persist:
             return self._persist_dataframe(
                 command.sql, conn, user_ns, append=False, index=not args.no_index
             )
-
+        elif args.persist_replace:
+            return self._persist_dataframe(
+                command.sql,
+                conn,
+                user_ns,
+                append=False,
+                index=not args.no_index,
+                replace=True,
+            )
         if args.append:
             return self._persist_dataframe(
                 command.sql, conn, user_ns, append=True, index=not args.no_index
@@ -402,6 +453,8 @@ class SqlMagic(Magics, Configurable):
             else:
                 if command.result_var:
                     self.shell.user_ns.update({command.result_var: result})
+                    if command.return_result_var:
+                        return result
                     return None
 
                 # Return results into the default ipython _ variable
@@ -410,39 +463,79 @@ class SqlMagic(Magics, Configurable):
         # JA: added DatabaseError for MySQL
         except (ProgrammingError, OperationalError, DatabaseError) as e:
             # Sqlite apparently return all errors as OperationalError :/
-
+            detailed_msg = detail(e, command.sql)
             if self.short_errors:
-                print(e)
+                if detailed_msg is not None:
+                    err = exceptions.UsageError(detailed_msg)
+                    raise err
+                else:
+                    print(e)
             else:
-                raise
+                if detailed_msg is not None:
+                    print(detailed_msg)
+                e.modify_exception = True
+                raise e
 
     legal_sql_identifier = re.compile(r"^[A-Za-z0-9#_$]+")
 
     @modify_exceptions
-    def _persist_dataframe(self, raw, conn, user_ns, append=False, index=True):
+    def _persist_dataframe(
+        self, raw, conn, user_ns, append=False, index=True, replace=False
+    ):
         """Implements PERSIST, which writes a DataFrame to the RDBMS"""
         if not DataFrame:
-            raise ImportError("Must `pip install pandas` to use DataFrames")
+            raise exceptions.MissingPackageError(
+                "You must install pandas to persist results: pip install pandas"
+            )
 
         frame_name = raw.strip(";")
 
-        # Get the DataFrame from the user namespace
+        # invalid identifier
+        if not frame_name.isidentifier():
+            raise exceptions.UsageError(
+                f"Expected {frame_name!r} to be a pd.DataFrame but it's"
+                " not a valid identifier"
+            )
+
+        # missing argument
         if not frame_name:
-            raise SyntaxError("Syntax: %sql --persist <name_of_data_frame>")
-        try:
-            frame = eval(frame_name, user_ns)
-        except SyntaxError:
-            raise SyntaxError("Syntax: %sql --persist <name_of_data_frame>")
+            raise exceptions.UsageError(
+                "Missing argument: %sql --persist <name_of_data_frame>"
+            )
+
+        # undefined variable
+        if frame_name not in user_ns:
+            raise exceptions.UsageError(
+                f"Expected {frame_name!r} to be a pd.DataFrame but it's undefined"
+            )
+
+        frame = user_ns[frame_name]
+
         if not isinstance(frame, DataFrame) and not isinstance(frame, Series):
-            raise TypeError("%s is not a Pandas DataFrame or Series" % frame_name)
+            raise exceptions.TypeError(
+                f"{frame_name!r} is not a Pandas DataFrame or Series"
+            )
 
         # Make a suitable name for the resulting database table
         table_name = frame_name.lower()
         table_name = self.legal_sql_identifier.search(table_name).group(0)
 
-        if_exists = "append" if append else "fail"
+        if replace:
+            if_exists = "replace"
+        elif append:
+            if_exists = "append"
+        else:
+            if_exists = "fail"
 
-        frame.to_sql(table_name, conn.session.engine, if_exists=if_exists, index=index)
+        try:
+            frame.to_sql(
+                table_name, conn.session.engine, if_exists=if_exists, index=index
+            )
+        except ValueError:
+            raise exceptions.ValueError(
+                f"""Table {table_name!r} already exists. Consider using \
+--persist-replace to drop the table before persisting the data frame"""
+            )
 
         display.message_success(f"Success! Persisted {table_name} to the database.")
 
@@ -461,6 +554,3 @@ def load_ipython_extension(ip):
     ip.register_magics(SqlCmdMagic)
 
     patch_ipython_usage_error(ip)
-
-
-# %%